added crud controllers for movies nad categories

+++ MovieLibrary.Core/Services/CategoryService.cs
@@ -16,46 +16,59 @@ namespace MovieLibrary.Core.Services
         public CategoryService(ICategoryRepository categoryRepository)
         {
             _categoryRepository = categoryRepository;
-
         }
-        IEnumerable<CategoryDto> IService<CategoryDto>.GetAll()
+
+        public IEnumerable<CategoryDto> GetAll() 
         {
             var categories = _categoryRepository.GetAll();
-
-            return categories.Select(category => new CategoryDto { 
+            return categories.Select(category => new CategoryDto
+            {
                 Id = category.Id,
                 Name = category.Name
             });
         }
-        CategoryDto IService<CategoryDto>.GetById(int id)
+
+        public CategoryDto GetById(int id) 
         {
             var category = _categoryRepository.GetById(id);
             if (category == null)
                 return null;
-            else return new CategoryDto { 
-                Id = category.Id,
-                Name = category.Name};
+            else
+                return new CategoryDto
+                {
+                    Id = category.Id,
+                    Name = category.Name
+                };
         }
 
-        void IService<CategoryDto>.Create(CategoryDto entity)
+        public CategoryDto Create(CategoryDto entity) 
         {
-            var category = new Category { 
-                Name = entity.Name};
-
+            var category = new Category
+            {
+                Name = entity.Name
+            };
             _categoryRepository.Add(category);
             entity.Id = category.Id;
+            return entity;
         }
-        void IService<CategoryDto>.Update(CategoryDto entity)
+
+        public CategoryDto Update(CategoryDto entity) 
         {
             var category = _categoryRepository.GetById(entity.Id);
-
-            category.Id = entity.Id;
+            if (category == null)
+                return null;
+            category.Name = entity.Name;
             _categoryRepository.Update(category);
+            return entity;
         }
-        public void Delete(int id)
+
+        public bool Delete(int id) 
         {
             var category = _categoryRepository.GetById(id);
+            if (category == null)
+                return false;
             _categoryRepository.Delete(category.Id);
+            return true;
         }
     }
 }
